Added test wasm

+++ core/src/action.rs
@@ -4,6 +4,6 @@ use common::address::Address;
 ///
 #[derive(Debug, Clone)]
 pub struct Action {
-    pub balance: u64,
+    pub balance: u32,
     pub addr: Address
 }
